refactor(ui): split ESP flash panel
Closes #2640
