+++ apps/ui/src/app/views/esp_flash_log_section.tsx
@@ -1,0 +1,14 @@
+import {
+  InlineEmptyState,
+  type EspFlashLogPanelModel,
+} from "./esp_flash_panel_shared";
+
+export function EspFlashLogContent(props: {
+  model: EspFlashLogPanelModel;
+}) {
+  const { model } = props;
+  if (model.emptyState) {
+    return <InlineEmptyState model={model.emptyState} />;
+  }
+  return <pre class="log-pre">{model.text}</pre>;
+}
