+++ apps/ui/src/app/views/esp_flash_readiness_section.tsx
@@ -1,0 +1,20 @@
+import {
+  MaintenanceNote,
+  StatusGrid,
+  type EspFlashReadinessPanelModel,
+} from "./esp_flash_panel_shared";
+
+export function EspFlashReadinessSection(props: {
+  model: EspFlashReadinessPanelModel;
+}) {
+  const { model } = props;
+  return (
+    <div class="maintenance-stack maintenance-stack--tight">
+      <div class="subtle">{model.summaryText}</div>
+      {model.rows.length > 0 ? <StatusGrid rows={model.rows} /> : null}
+      {model.errorText ? (
+        <MaintenanceNote text={model.errorText} variant="bad" />
+      ) : null}
+    </div>
+  );
+}
